Wrap PhotonNetwork.LoadLevel with MasterClient check, allow Player & Grabbable layers to interact via physics settings
In NetworkLauncher, we need to only call PhotonNetwork if MasterClient. We were calling it on everyone, which can do extra scene loads. I'm guessing this contributed to the stuttering. Here's what that looks like now

Changed player layer to be interactable with grabbables in the project settings > physics section. I think we could grab gun before because it was in a snap zone, the hammers are not, so we need to be able to interact with them (they have grabbable layer)